pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Michael Agranovskiy <csagranovskiy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
mysqlclient = "^2.1.0"
django-model-utils = "^4.2.0"
oauth2client = "^4.1.3"
requests = "^2.27.1"
ipython = "^8.0.0"
pyOpenSSL = "^21.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
